Coverity CID 468157
(cherry picked from commit b7a94bdb0a21dee121b30516492c58dfe17418a9)

+++ vnc/vnc.c
@@ -2094,7 +2094,7 @@ lib_mod_set_param(struct vnc *v, const char *name, const char *value)
     }
     else if (g_strcasecmp(name, "disabled_encodings_mask") == 0)
     {
-        v->enabled_encodings_mask = ~g_atoi(value);
+        v->enabled_encodings_mask = (unsigned int)~g_atoi(value);
     }
     else if (g_strcasecmp(name, "client_info") == 0)
     {
